Terrible code still with a bug

+++ src/main/java/ca/uqam/example/model/passenger/BusinessClassPassenger.java
@@ -2,7 +2,7 @@ package ca.uqam.example.model.passenger;
 
 public class BusinessClassPassenger extends Passenger {
 
-    public BusinessClassPassenger(String passport, String name, double age) {
+    public BusinessClassPassenger(String passport, String name, int age) {
         super(passport, name, age);
     }
