+++ src/main/java/ca/uqam/example/model/passenger/FirstClassPassenger.java
@@ -2,7 +2,7 @@ package ca.uqam.example.model.passenger;
 
 public class FirstClassPassenger extends Passenger {
 
-    public FirstClassPassenger(String passport, String name, double age) {
+    public FirstClassPassenger(String passport, String name, int age) {
         super(passport, name, age);
     }
